Fix arm menu needing to be open for commands to work
