+++ Unity/Assets/Scripts/Components/Elements/HololensCommandManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f9a2c4e8d5b4c6fa1e72b9c0d3e4f57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
